Removed hard coding of service name in clean up script.
Changed README file to alert user not to add services ahead of already defined (base) service or the Cleanup.sh script will break.  Now reading the name of the first service in the docker-compose.yml file, which spares the user the effort of having to change Cleanup.sh if the service name is changed.